[shell] Move mouse-related types to mouse.rs
This will likely break some existing code; to make that less painful
I've re-exported these types from their previous location, attaching
deprecation warnings pointing to the new location.
