+++ druid/src/lib.rs
@@ -49,7 +49,8 @@ pub use unicode_segmentation;
 pub use druid_shell::clipboard::ClipboardItem;
 pub use druid_shell::dialog::{FileDialogOptions, FileDialogType};
 pub use druid_shell::keyboard::{KeyCode, KeyEvent, KeyModifiers};
-pub use druid_shell::window::{Cursor, MouseButton, TimerToken};
+pub use druid_shell::mouse::{Cursor, MouseButton};
+pub use druid_shell::window::TimerToken;
 use druid_shell::window::{Text, WinCtx, WindowHandle};
 pub use shell::hotkey::{HotKey, RawMods, SysMods};
