+++ druid/src/win_handler.rs
@@ -26,7 +26,8 @@ use crate::kurbo::{Rect, Size, Vec2};
 use crate::piet::{Piet, RenderContext};
 use crate::shell::application::Application;
 use crate::shell::dialog::FileDialogOptions;
-use crate::shell::window::{Cursor, MouseEvent, WinCtx, WinHandler, WindowHandle};
+use crate::shell::mouse::{Cursor, MouseEvent};
+use crate::shell::window::{WinCtx, WinHandler, WindowHandle};
 
 use crate::app_delegate::{AppDelegate, DelegateCtx};
 use crate::menu::ContextMenu;
